Add support for non-45 degree angles
The Ball class was heavily changed so that it could move in non-45 degree angles. How it was able to do that was mostly taken from Arduboy2's implementation of Ardyboy2Base::drawLine, which came from Bresenham's line algorithm. For now you can change the velocity of the ball in game.cpp for whatever kind of angle you want, not change the angle based on where it hit the ball.