Add paddle reflections
When the ball hits the paddle, it gets a new velocity based on where it hit the paddle. This can be done by finding the difference between the center of the paddle and the center of the ball (because the width of the ball is just 2, its position is used instead).