Add 2 player support
The second player is able to play by pressing B to go up and A to go down. However, the A button is also used as the pause button.

To get around this, the pause button was moved to the left arrow button instead.

You can select if you want to play in 1player or 2player mode in the title screen.