Change paddle terminology
Previously, the left paddle was called Player and the right paddle was called Computer. This was inconsistent with 2 player mode, where both paddles aren't controlled by the computer. To fix this, they were renamed to Player1 and Player2, and the win/lose messages were reworded to indicate that either player 1 or player 2 won rather than just player 1 winning or losing.